Collimation infrastructure, to be tested
