+++ examples/secondary_production/000_secondary_production.py
@@ -49,43 +49,7 @@ class DummyInteractionProcess:
 
         return products
 
-class BeamInteraction:
-
-    def __init__(self, interaction_process):
-        self.interaction_process = interaction_process
-
-    def track(self, particles):
-
-        assert isinstance(particles._buffer.context, xo.ContextCpu)
-        assert particles._num_active_particles >= 0
-
-        # Assumes active particles are contiguous
-        products = self.interaction_process.interact(particles)
-
-        # TODO: This should work also when no products are there
-        #       Particles reorganization should still happen
-
-        if products is None or products['x'].size == 0:
-            particles.reorganize()
-        else:
-            new_particles = xt.Particles(_context=particles._buffer.context,
-                    p0c = particles.p0c[0], # TODO: Should we check that 
-                                            #       they are all the same?
-                    s = products['s'],
-                    x = products['x'],
-                    px = products['px'],
-                    y = products['y'],
-                    py = products['py'],
-                    zeta = products['zeta'],
-                    delta = products['delta'],
-                    mass_ratio = products['mass_ratio'],
-                    charge_ratio = products['charge_ratio'],
-                    parent_id = products['parent_id'])
-
-            particles.add_particles(new_particles)
-
-
-beam_interaction = BeamInteraction(
+beam_interaction = xt.BeamInteraction(
         interaction_process=DummyInteractionProcess(fraction_lost=0.1,
                                                     fraction_secondary=0.2))
