+++ xtrack/beam_elements/__init__.py
@@ -1,2 +1,3 @@
 from .elements import *
 from .apertures import *
+from .beam_interaction import BeamInteraction
